+++ changelog.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Exclude filter.
 
+### Removed
+- $exclude parameter from Pattern filter.
+
 ## [0.1.1] - 2019-01-15
 ### Added
 - Stream logger.